+++ packages/web3-providers/src/SmartPay/apis/AbstractAccountAPI.ts
@@ -20,8 +20,8 @@ import { isSameAddress } from '@masknet/web3-shared-base'
 import WalletABI from '@masknet/web3-contracts/abis/Wallet.json'
 import type { Wallet } from '@masknet/web3-contracts/types/Wallet.js'
 import { LOG_ROOT, MAX_ACCOUNT_LENGTH } from '../constants.js'
-import { SmartPayBundlerAPI } from './SmartPayBundlerAPI.js'
-import { SmartPayFunderAPI } from './SmartPayFunderAPI.js'
+import { SmartPayBundlerAPI } from './BundlerAPI.js'
+import { SmartPayFunderAPI } from './FunderAPI.js'
 import { MulticallAPI } from '../../Multicall/index.js'
 import { Web3API } from '../../EVM/index.js'
 import type { AbstractAccountAPI } from '../../entry-types.js'
